requirements.txt
asgiref==3.3.4
Django==3.2.3
django-redis==4.12.1
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
